rerouted and fixed up components and pages
