+++ src/pages/InvitePage.jsx
@@ -1,0 +1,16 @@
+import "../styles.css";
+import Header from "../components/Header.jsx";
+import Footer from "../components/Footer.jsx";
+import InviteForm from "../components/InviteForm.jsx"
+
+function InvitePage() {
+    return (
+    <>
+        <Header />
+        <InviteForm />
+        <Footer />;
+    </>
+    );
+}
+
+export default InvitePage;
